requirements.txt
python-telegram-bot==20.7
rapidfuzz==3.6.1
